refactor: remove service interface
