+++ src/main/java/com/example/userservice/controller/ClientController.java
@@ -3,14 +3,10 @@ package com.example.userservice.controller;
 import java.util.List;
 import java.util.Optional;
 
-import com.example.userservice.dto.EmailDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.example.userservice.dto.ClientDTO;
 import com.example.userservice.mapper.ClientAdapter;
@@ -46,5 +42,6 @@ public class ClientController {
 	public ResponseEntity<?> getClientEmailAndName(@PathVariable Long cartId) {
 		return clientService.getClientEmailAndName(cartId);
 	}
+
 }
